move visual thinker definition into its own header
